+++ src/Cwaff-Guns/Femtobyte.cs
@@ -158,7 +158,8 @@ public class Femtobyte : CwaffGun
           .InitProjectile(GunData.New(clipSize: -1, angleVariance: 2.0f, shootStyle: ShootStyle.SemiAutomatic, damage: 7.5f, speed: 90.0f, hideAmmo: true,
             cooldown: 0.4f, sprite: "femtobyte_projectile", fps: 2, anchor: Anchor.MiddleCenter, hitEnemySound: "femtobyte_hit_enemy_sound"))
           .Attach<FemtobyteProjectile>()
-          .AttachTrail("femtobyte_beam", fps: 10, cascadeTimer: C.FRAME, softMaxLength: 1f, destroyOnEmpty: false);
+          .AttachTrail("femtobyte_beam", fps: 10, cascadeTimer: C.FRAME, softMaxLength: 1f, destroyOnEmpty: false,
+            dispersalPrefab: Lazy.DispersalParticles(Color.white));
 
         _ImpactBits = VFX.Create("femtobyte_projectile_vfx");
     }